server/requirements.txt
bcrypt==4.1.3
bidict==0.23.1
blinker==1.8.2
click==8.1.7
dnspython==2.6.1
eventlet==0.36.1
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Mail==0.9.1
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
gevent==24.2.1
greenlet==3.0.3
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
PyJWT==2.8.0
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.2

# ------ PostgresSQL dependencies ----------
# psycopg2-binary
# psycopg2
# ------ END PostgresSQL dependencies ------

simple-websocket==1.0.0
SQLAlchemy==2.0.30
typing_extensions==4.11.0
Werkzeug==3.0.3
wsproto==1.2.0
zope.event==5.0
zope.interface==6.3
